feat: Add Auction Bounded Context

+++ src/main/java/com/auction/auction/security/model/Customer.java
@@ -53,5 +53,5 @@ public class Customer {
     private Long bonus;
 
     @Column(name = "wallet")
-    private Long wallet;
+    private Double wallet;
 }
